No more warning
git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@99091 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: dd414f580c388f3f7b1b7efb3710f9d3491b0fcf

+++ Online/Online/GaudiOnline/src/SLICEReaderSvc.cpp
@@ -343,7 +343,7 @@ StatusCode SLICEReaderSvc::run()
 
     do
     {
-        for(m_CurFile = 0; m_CurFile < m_InputFiles.size() && m_ReaderState==RUNNING && sc.isSuccess(); ++m_CurFile) {
+        for(m_CurFile = 0; (unsigned int) m_CurFile < m_InputFiles.size() && m_ReaderState==RUNNING && sc.isSuccess(); ++m_CurFile) {
             sc=readFile();
         }
     }
